heaps of stuff and a nasty one

+++ src/swmmwriteandread.h
@@ -37,7 +37,7 @@ class SWMMWriteAndRead
 public:
 	SWMMWriteAndRead(std::map<string, DM::ViewContainer *>, std::string rainfile, std::string filename  = "");
 	void setRainFile(std::string rainfile);
-	void setClimateChangeFactor(int cf);
+	void setClimateChangeFactor(double cf);
 	void writeSWMMFile();
 	void setupSWMM();
 	std::string getSWMMUUIDPath();
